Move shares state machines to shares_base
