+++ backend/dataall/modules/s3_datasets_shares/db/share_object_repositories.py
@@ -12,303 +12,18 @@ from dataall.core.organizations.db.organization_models import Organization
 from dataall.base.db import exceptions, paginate
 from dataall.modules.shares_base.services.shares_enums import (
     ShareItemHealthStatus,
-    ShareObjectActions,
     ShareObjectStatus,
-    ShareItemActions,
     ShareItemStatus,
     ShareableType,
     PrincipalType,
 )
+from dataall.modules.shares_base.db.share_object_state_machines import ShareItemSM
 from dataall.modules.s3_datasets_shares.db.share_object_models import ShareObjectItem, ShareObject
 from dataall.modules.s3_datasets.db.dataset_repositories import DatasetRepository
 from dataall.modules.s3_datasets.db.dataset_models import DatasetStorageLocation, DatasetTable, S3Dataset, DatasetBucket
 from dataall.modules.datasets_base.db.dataset_models import DatasetBase
 
 logger = logging.getLogger(__name__)
-
-
-class Transition:
-    def __init__(self, name, transitions):
-        self._name = name
-        self._transitions = transitions
-        self._all_source_states = [*set([item for sublist in transitions.values() for item in sublist])]
-        self._all_target_states = [item for item in transitions.keys()]
-
-    def validate_transition(self, prev_state):
-        if prev_state in self._all_target_states:
-            logger.info(f'Resource is already in target state ({prev_state}) in {self._all_target_states}')
-            return False
-        elif prev_state not in self._all_source_states:
-            raise exceptions.UnauthorizedOperation(
-                action=self._name,
-                message=f'This transition is not possible, {prev_state} cannot go to {self._all_target_states}. '
-                f'If there is a sharing or revoking in progress wait until it is complete and try again.',
-            )
-        else:
-            return True
-
-    def get_transition_target(self, prev_state):
-        if self.validate_transition(prev_state):
-            for target_state, list_prev_states in self._transitions.items():
-                if prev_state in list_prev_states:
-                    return target_state
-                else:
-                    pass
-        else:
-            return prev_state
-
-
-class ShareObjectSM:
-    def __init__(self, state):
-        self._state = state
-        self.transitionTable = {
-            ShareObjectActions.Submit.value: Transition(
-                name=ShareObjectActions.Submit.value,
-                transitions={
-                    ShareObjectStatus.Submitted.value: [ShareObjectStatus.Draft.value, ShareObjectStatus.Rejected.value]
-                },
-            ),
-            ShareObjectActions.Approve.value: Transition(
-                name=ShareObjectActions.Approve.value,
-                transitions={ShareObjectStatus.Approved.value: [ShareObjectStatus.Submitted.value]},
-            ),
-            ShareObjectActions.Reject.value: Transition(
-                name=ShareObjectActions.Reject.value,
-                transitions={ShareObjectStatus.Rejected.value: [ShareObjectStatus.Submitted.value]},
-            ),
-            ShareObjectActions.RevokeItems.value: Transition(
-                name=ShareObjectActions.RevokeItems.value,
-                transitions={
-                    ShareObjectStatus.Revoked.value: [
-                        ShareObjectStatus.Draft.value,
-                        ShareObjectStatus.Submitted.value,
-                        ShareObjectStatus.Rejected.value,
-                        ShareObjectStatus.Processed.value,
-                    ]
-                },
-            ),
-            ShareObjectActions.Start.value: Transition(
-                name=ShareObjectActions.Start.value,
-                transitions={
-                    ShareObjectStatus.Share_In_Progress.value: [ShareObjectStatus.Approved.value],
-                    ShareObjectStatus.Revoke_In_Progress.value: [ShareObjectStatus.Revoked.value],
-                },
-            ),
-            ShareObjectActions.Finish.value: Transition(
-                name=ShareObjectActions.Finish.value,
-                transitions={
-                    ShareObjectStatus.Processed.value: [
-                        ShareObjectStatus.Share_In_Progress.value,
-                        ShareObjectStatus.Revoke_In_Progress.value,
-                    ],
-                },
-            ),
-            ShareObjectActions.FinishPending.value: Transition(
-                name=ShareObjectActions.FinishPending.value,
-                transitions={
-                    ShareObjectStatus.Draft.value: [
-                        ShareObjectStatus.Revoke_In_Progress.value,
-                    ],
-                },
-            ),
-            ShareObjectActions.Delete.value: Transition(
-                name=ShareObjectActions.Delete.value,
-                transitions={
-                    ShareObjectStatus.Deleted.value: [
-                        ShareObjectStatus.Rejected.value,
-                        ShareObjectStatus.Draft.value,
-                        ShareObjectStatus.Submitted.value,
-                        ShareObjectStatus.Processed.value,
-                    ]
-                },
-            ),
-            ShareItemActions.AddItem.value: Transition(
-                name=ShareItemActions.AddItem.value,
-                transitions={
-                    ShareObjectStatus.Draft.value: [
-                        ShareObjectStatus.Submitted.value,
-                        ShareObjectStatus.Rejected.value,
-                        ShareObjectStatus.Processed.value,
-                    ]
-                },
-            ),
-            ShareObjectActions.AcquireLockFailure.value: Transition(
-                name=ShareObjectActions.AcquireLockFailure.value,
-                transitions={
-                    ShareObjectStatus.Processed.value: [
-                        ShareObjectStatus.Share_In_Progress.value,
-                        ShareObjectStatus.Revoke_In_Progress.value,
-                    ]
-                },
-            ),
-        }
-
-    def run_transition(self, transition):
-        trans = self.transitionTable[transition]
-        new_state = trans.get_transition_target(self._state)
-        return new_state
-
-    def update_state(self, session, share, new_state):
-        logger.info(f'Updating share object {share.shareUri} in DB from {self._state} to state {new_state}')
-        ShareObjectRepository.update_share_object_status(session=session, share_uri=share.shareUri, status=new_state)
-        self._state = new_state
-        return True
-
-
-class ShareItemSM:
-    def __init__(self, state):
-        self._state = state
-        self.transitionTable = {
-            ShareItemActions.AddItem.value: Transition(
-                name=ShareItemActions.AddItem.value,
-                transitions={ShareItemStatus.PendingApproval.value: [ShareItemStatus.Deleted.value]},
-            ),
-            ShareObjectActions.Submit.value: Transition(
-                name=ShareObjectActions.Submit.value,
-                transitions={
-                    ShareItemStatus.PendingApproval.value: [
-                        ShareItemStatus.Share_Rejected.value,
-                        ShareItemStatus.Share_Failed.value,
-                    ],
-                    ShareItemStatus.Revoke_Approved.value: [ShareItemStatus.Revoke_Approved.value],
-                    ShareItemStatus.Revoke_Failed.value: [ShareItemStatus.Revoke_Failed.value],
-                    ShareItemStatus.Share_Approved.value: [ShareItemStatus.Share_Approved.value],
-                    ShareItemStatus.Share_Succeeded.value: [ShareItemStatus.Share_Succeeded.value],
-                    ShareItemStatus.Revoke_Succeeded.value: [ShareItemStatus.Revoke_Succeeded.value],
-                    ShareItemStatus.Share_In_Progress.value: [ShareItemStatus.Share_In_Progress.value],
-                    ShareItemStatus.Revoke_In_Progress.value: [ShareItemStatus.Revoke_In_Progress.value],
-                },
-            ),
-            ShareObjectActions.Approve.value: Transition(
-                name=ShareObjectActions.Approve.value,
-                transitions={
-                    ShareItemStatus.Share_Approved.value: [ShareItemStatus.PendingApproval.value],
-                    ShareItemStatus.Revoke_Approved.value: [ShareItemStatus.Revoke_Approved.value],
-                    ShareItemStatus.Revoke_Failed.value: [ShareItemStatus.Revoke_Failed.value],
-                    ShareItemStatus.Share_Succeeded.value: [ShareItemStatus.Share_Succeeded.value],
-                    ShareItemStatus.Revoke_Succeeded.value: [ShareItemStatus.Revoke_Succeeded.value],
-                    ShareItemStatus.Share_In_Progress.value: [ShareItemStatus.Share_In_Progress.value],
-                    ShareItemStatus.Revoke_In_Progress.value: [ShareItemStatus.Revoke_In_Progress.value],
-                },
-            ),
-            ShareObjectActions.Reject.value: Transition(
-                name=ShareObjectActions.Reject.value,
-                transitions={
-                    ShareItemStatus.Share_Rejected.value: [ShareItemStatus.PendingApproval.value],
-                    ShareItemStatus.Revoke_Approved.value: [ShareItemStatus.Revoke_Approved.value],
-                    ShareItemStatus.Revoke_Failed.value: [ShareItemStatus.Revoke_Failed.value],
-                    ShareItemStatus.Share_Succeeded.value: [ShareItemStatus.Share_Succeeded.value],
-                    ShareItemStatus.Revoke_Succeeded.value: [ShareItemStatus.Revoke_Succeeded.value],
-                    ShareItemStatus.Share_In_Progress.value: [ShareItemStatus.Share_In_Progress.value],
-                    ShareItemStatus.Revoke_In_Progress.value: [ShareItemStatus.Revoke_In_Progress.value],
-                },
-            ),
-            ShareObjectActions.Start.value: Transition(
-                name=ShareObjectActions.Start.value,
-                transitions={
-                    ShareItemStatus.Share_In_Progress.value: [ShareItemStatus.Share_Approved.value],
-                    ShareItemStatus.Revoke_In_Progress.value: [ShareItemStatus.Revoke_Approved.value],
-                },
-            ),
-            ShareItemActions.Success.value: Transition(
-                name=ShareItemActions.Success.value,
-                transitions={
-                    ShareItemStatus.Share_Succeeded.value: [ShareItemStatus.Share_In_Progress.value],
-                    ShareItemStatus.Revoke_Succeeded.value: [ShareItemStatus.Revoke_In_Progress.value],
-                },
-            ),
-            ShareItemActions.Failure.value: Transition(
-                name=ShareItemActions.Failure.value,
-                transitions={
-                    ShareItemStatus.Share_Failed.value: [ShareItemStatus.Share_In_Progress.value],
-                    ShareItemStatus.Revoke_Failed.value: [ShareItemStatus.Revoke_In_Progress.value],
-                },
-            ),
-            ShareItemActions.RemoveItem.value: Transition(
-                name=ShareItemActions.RemoveItem.value,
-                transitions={
-                    ShareItemStatus.Deleted.value: [
-                        ShareItemStatus.PendingApproval.value,
-                        ShareItemStatus.Share_Rejected.value,
-                        ShareItemStatus.Share_Failed.value,
-                        ShareItemStatus.Revoke_Succeeded.value,
-                    ]
-                },
-            ),
-            ShareObjectActions.RevokeItems.value: Transition(
-                name=ShareObjectActions.RevokeItems.value,
-                transitions={
-                    ShareItemStatus.Revoke_Approved.value: [
-                        ShareItemStatus.Share_Succeeded.value,
-                        ShareItemStatus.Revoke_Failed.value,
-                        ShareItemStatus.Revoke_Approved.value,
-                    ]
-                },
-            ),
-            ShareObjectActions.Delete.value: Transition(
-                name=ShareObjectActions.Delete.value,
-                transitions={
-                    ShareItemStatus.Deleted.value: [
-                        ShareItemStatus.PendingApproval.value,
-                        ShareItemStatus.Share_Rejected.value,
-                        ShareItemStatus.Share_Failed.value,
-                        ShareItemStatus.Revoke_Succeeded.value,
-                    ]
-                },
-            ),
-            ShareObjectActions.AcquireLockFailure.value: Transition(
-                name=ShareObjectActions.AcquireLockFailure.value,
-                transitions={
-                    ShareItemStatus.Share_Failed.value: [ShareItemStatus.Share_Approved.value],
-                    ShareItemStatus.Revoke_Failed.value: [ShareItemStatus.Revoke_Approved.value],
-                },
-            ),
-        }
-
-    def run_transition(self, transition):
-        trans = self.transitionTable[transition]
-        new_state = trans.get_transition_target(self._state)
-        return new_state
-
-    def update_state(self, session, share_uri, new_state):
-        if share_uri and (new_state != self._state):
-            if new_state == ShareItemStatus.Deleted.value:
-                logger.info(f'Deleting share items in DB in {self._state} state')
-                ShareObjectRepository.delete_share_item_status_batch(
-                    session=session, share_uri=share_uri, status=self._state
-                )
-            else:
-                logger.info(f'Updating share items in DB from {self._state} to state {new_state}')
-                ShareObjectRepository.update_share_item_status_batch(
-                    session=session, share_uri=share_uri, old_status=self._state, new_status=new_state
-                )
-            self._state = new_state
-        else:
-            logger.info(f'Share Items in DB already in target state {new_state} or no update is required')
-            return True
-
-    def update_state_single_item(self, session, share_item, new_state):
-        logger.info(f'Updating share item in DB {share_item.shareItemUri} status to {new_state}')
-        ShareObjectRepository.update_share_item_status(session=session, uri=share_item.shareItemUri, status=new_state)
-        self._state = new_state
-        return True
-
-    @staticmethod
-    def get_share_item_shared_states():
-        return [
-            ShareItemStatus.Share_Succeeded.value,
-            ShareItemStatus.Share_In_Progress.value,
-            ShareItemStatus.Revoke_Failed.value,
-            ShareItemStatus.Revoke_In_Progress.value,
-            ShareItemStatus.Revoke_Approved.value,
-        ]
-
-    @staticmethod
-    def get_share_item_revokable_states():
-        return [
-            ShareItemStatus.Share_Succeeded.value,
-            ShareItemStatus.Revoke_Failed.value,
-        ]
 
 
 class ShareEnvironmentResource(EnvironmentResource):
@@ -362,7 +77,7 @@ class ShareObjectRepository:
             return session.query(DatasetBucket).get(item_uri)
 
     @staticmethod
-    def get_share_by_uri(session, uri):
+    def get_share_by_uri(session, uri):  ## TODO: Already in shares_base
         share = session.query(ShareObject).get(uri)
         if not share:
             raise exceptions.ObjectNotFound('Share', uri)
@@ -689,7 +404,9 @@ class ShareObjectRepository:
         return share
 
     @staticmethod
-    def update_share_object_status(session, share_uri: str, status: str) -> ShareObject:
+    def update_share_object_status(
+        session, share_uri: str, status: str
+    ) -> ShareObject:  ## TODO: Already in shares_base
         share = ShareObjectRepository.get_share_by_uri(session, share_uri)
         share.status = status
         session.commit()
@@ -725,7 +442,7 @@ class ShareObjectRepository:
         session,
         share_uri: str,
         status: str,
-    ):
+    ):  ## TODO: Already in shares_base
         (
             session.query(ShareObjectItem)
             .filter(and_(ShareObjectItem.shareUri == share_uri, ShareObjectItem.status == status))
@@ -756,7 +473,7 @@ class ShareObjectRepository:
         share_uri: str,
         old_status: str,
         new_status: str,
-    ) -> bool:
+    ) -> bool:  ## TODO: Already in shares_base
         (
             session.query(ShareObjectItem)
             .filter(and_(ShareObjectItem.shareUri == share_uri, ShareObjectItem.status == old_status))
