+++ backend/dataall/modules/s3_datasets_shares/services/share_managers/lf_share_manager.py
@@ -11,7 +11,8 @@ from dataall.base.aws.quicksight import QuicksightClient
 from dataall.base.aws.iam import IAM
 from dataall.base.aws.sts import SessionHelper
 from dataall.base.db import exceptions
-from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository, ShareItemSM
+from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository
+from dataall.modules.shares_base.db.share_object_state_machines import ShareItemSM
 from dataall.modules.shares_base.services.shares_enums import (
     ShareItemStatus,
     ShareObjectActions,
