+++ backend/dataall/modules/s3_datasets_shares/services/share_object_service.py
@@ -20,8 +20,8 @@ from dataall.modules.shares_base.services.shares_enums import (
     PrincipalType,
 )
 from dataall.modules.s3_datasets_shares.db.share_object_models import ShareObjectItem, ShareObject
-from dataall.modules.s3_datasets_shares.db.share_object_repositories import (
-    ShareObjectRepository,
+from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository
+from dataall.modules.shares_base.db.share_object_state_machines import (
     ShareObjectSM,
     ShareItemSM,
 )
