+++ backend/dataall/modules/s3_datasets_shares/services/share_processors/s3_access_point_process_share.py
@@ -13,7 +13,8 @@ from dataall.modules.shares_base.services.shares_enums import (
     ShareItemActions,
 )
 from dataall.modules.s3_datasets_shares.db.share_object_models import ShareObject
-from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository, ShareItemSM
+from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository
+from dataall.modules.shares_base.db.share_object_state_machines import ShareItemSM
 
 
 log = logging.getLogger(__name__)
